requirements.txt
requests
web3
pytest
fastapi
uvicorn
httpx
python-dotenv
colorama